Update README.md
Important: To add functionality to more websites, the permission "all_urls" was necessary, however with the recent update to manifest v3, it's optional. If the extension can't acceess audio on a ceertain site do the following: Right click the extension, click "manage extension", click the permission tab, allow "Access your data for all websites". This is a necessary permission to acceess iframes. The extension is open source and fully adheres to this repository. There's no form of analytics in this extension. So don't be afraid fo the permissions.